:bug: Fix redirections
The fix for this was actually pretty involved, as the previous redirection code worked (but actually didn't.) The fix was to just spin up a minimal express server on HTTP just for redirections to the HTTPS server, and it seems to be working in my 30 seconds of testing. Let's hope it actually does work, though.