:sparkles: ingenious arbitary endpoint injection
Instead of manually modifying each method call to the Axios instance, we intercept that call and modify the method--replacing the /admin part with the appropriate arbitrary endpoint. We cache the original get and post methods and bind them to the Axios instance, passing in the modified URL argument, allowing for a seamless change without breaking DRY. Feeling proud about this one.